# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "SimbaTracking",
        "description": "SimbaTracking"
    },
    "paths": {
        "/customer/signin": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "SignIn",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "SignIn",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "device_token": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.SignIn"
            }
        },
        "/customer/forgot": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "ForgotPassword",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ForgotPassword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.ForgotPassword"
            }
        },
        "/customer/signup": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "SignUp",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "SignUp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "device_token": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.SignUp"
            }
        },
        "/customer/get-profile": {
            "get": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "GetProfile",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "customer_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "CustomerController.GetProfile"
            }
        },
        "/customer/update-profile": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "UpdateProfile",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UpdateProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "customer_type": {
                                        "type": "string",
                                        "description": "individual or business"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    },
                                    "business_mobile_country_code": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    },
                                    "business_mobile": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    },
                                    "preferred_contact": {
                                        "type": "string",
                                        "description": "Pass while customer_type=business"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.UpdateProfile"
            }
        },
        "/customer/change-password": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "ChangePassword",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ChangePassword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.ChangePassword"
            }
        },
        "/customer/logout": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "Logout",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.Logout"
            }
        },
        "/customer/create-job": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "CreateJob",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CreateJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "truck_id": {
                                        "type": "string"
                                    },
                                    "delivery_category_id": {
                                        "type": "string"
                                    },
                                    "pickup_from": {
                                        "type": "string",
                                        "description": "YYYY-MM-DD HH:mm"
                                    },
                                    "pickup_landmark": {
                                        "type": "string"
                                    },
                                    "pickup_address": {
                                        "type": "string"
                                    },
                                    "pickup_latitude": {
                                        "type": "number"
                                    },
                                    "pickup_longitude": {
                                        "type": "number"
                                    },
                                    "delivery_landmark": {
                                        "type": "string"
                                    },
                                    "delivery_address": {
                                        "type": "string"
                                    },
                                    "delivery_latitude": {
                                        "type": "number"
                                    },
                                    "delivery_longitude": {
                                        "type": "number"
                                    },
                                    "pickup_contact_name": {
                                        "type": "string"
                                    },
                                    "pickup_contact_number": {
                                        "type": "string"
                                    },
                                    "pickup_instructions": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "string",
                                        "description": "Pass object as a string => {item_id:'ITM24234', item_title:'dads', volume:12, weight:12, description:'dfgdfg', qty:10}"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.CreateJob"
            }
        },
        "/customer/update-job": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "UpdateJob",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "UpdateJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "job_id": {
                                        "type": "string"
                                    },
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "truck_id": {
                                        "type": "string"
                                    },
                                    "delivery_category_id": {
                                        "type": "string"
                                    },
                                    "pickup_from": {
                                        "type": "string",
                                        "description": "YYYY-MM-DD HH:mm"
                                    },
                                    "pickup_landmark": {
                                        "type": "string"
                                    },
                                    "pickup_address": {
                                        "type": "string"
                                    },
                                    "pickup_latitude": {
                                        "type": "number"
                                    },
                                    "pickup_longitude": {
                                        "type": "number"
                                    },
                                    "delivery_landmark": {
                                        "type": "string"
                                    },
                                    "delivery_address": {
                                        "type": "string"
                                    },
                                    "delivery_latitude": {
                                        "type": "number"
                                    },
                                    "delivery_longitude": {
                                        "type": "number"
                                    },
                                    "pickup_contact_name": {
                                        "type": "string"
                                    },
                                    "pickup_contact_number": {
                                        "type": "string"
                                    },
                                    "pickup_instructions": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "string",
                                        "description": "Pass object as a string => {item_id:'ITM24234', volume:12, weight:12, description:'dfgdfg', qty:10}"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.UpdateJob"
            }
        },
...
        "/customer/rate-job": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "RateJob",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "RateJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "description": "Optional"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.RateJob"
            }
        },
...
    },
...
    "servers": [{
        "url": "http://localhost:1330/api"
    }, {
        "url": "http://15.188.211.169:1330/api"
    }]
}